requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alkivi-logger==1.2.4
alkivi-odoo-client==1.0.3
click==6.7
odoorpc==0.6.0            # via alkivi-odoo-client
